fichier.c

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: fsidler <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 15:54:40 by fsidler           #+#    #+#             */
-/*   Updated: 2015/12/01 13:57:30 by fsidler          ###   ########.fr       */
+/*   Updated: 2015/12/02 18:10:07 by fsidler          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *str, int c)
 {
-	int	i;
+	long long	i;
 
 	i = 0;
 	while (str[i])
